fixed issue in gen obs

+++ fantasy_chess/env/gameClasses.py
@@ -292,10 +292,10 @@ class GameManager():
                 c2 = 0*self.board.units_map
                 c3 = 0*self.board.units_map
                 for opp in opponent:
-                    c2 = c2 + self.board.units_map == opp
+                    c2 = c2 + (self.board.units_map == opp).astype(np.int8)
 
                 for friend in friendly:
-                    c3 = c3 + self.board.units_map == friend
+                    c3 = c3 + (self.board.units_map == friend).astype(np.int8)
 
                 channels.append(c2)
                 channels.append(c3)
